typings cleanup
