+++ tabby-core/src/api/index.ts
@@ -6,7 +6,7 @@ export { TabRecoveryProvider, RecoveryToken } from './tabRecovery'
 export { ToolbarButtonProvider, ToolbarButton } from './toolbarButtonProvider'
 export { ConfigProvider } from './configProvider'
 export { HotkeyProvider, HotkeyDescription, Hotkey } from './hotkeyProvider'
-export { Theme } from './theme'
+export { Theme, TerminalColorScheme } from './theme'
 export { TabContextMenuItemProvider } from './tabContextMenuProvider'
 export { SelectorOption } from './selector'
 export { CLIHandler, CLIEvent } from './cli'
