+++ tabby-terminal/src/api/interfaces.ts
@@ -1,19 +1,8 @@
-import { ConnectableProfile, Profile } from 'tabby-core'
+import { ConnectableProfile, Profile, TerminalColorScheme } from 'tabby-core'
 
 export interface ResizeEvent {
     columns: number
     rows: number
-}
-
-export interface TerminalColorScheme {
-    name: string
-    foreground: string
-    background: string
-    cursor: string
-    colors: string[]
-    selection?: string
-    selectionForeground?: string
-    cursorAccent?: string
 }
 
 export interface BaseTerminalProfile extends Profile {
