+++ tabby-terminal/src/components/colorSchemeSelector.component.ts
@@ -2,9 +2,8 @@
 import { marker as _ } from '@biesbjerg/ngx-translate-extract-marker'
 
 import { Component, Inject, Input, ChangeDetectionStrategy, ChangeDetectorRef, HostBinding, Output, EventEmitter } from '@angular/core'
-import { ConfigService } from 'tabby-core'
+import { ConfigService, TerminalColorScheme } from 'tabby-core'
 import { TerminalColorSchemeProvider } from '../api/colorSchemeProvider'
-import { TerminalColorScheme } from '../api/interfaces'
 
 _('Search color schemes')
