+++ tabby-terminal/src/components/colorSchemeSettingsForMode.component.ts
@@ -3,9 +3,8 @@ import deepEqual from 'deep-equal'
 import { marker as _ } from '@biesbjerg/ngx-translate-extract-marker'
 
 import { Component, Inject, Input, ChangeDetectionStrategy, ChangeDetectorRef, HostBinding } from '@angular/core'
-import { ConfigService, PlatformService, TranslateService } from 'tabby-core'
+import { ConfigService, PlatformService, TerminalColorScheme, TranslateService } from 'tabby-core'
 import { TerminalColorSchemeProvider } from '../api/colorSchemeProvider'
-import { TerminalColorScheme } from '../api/interfaces'
 
 _('Search color schemes')
